Sort of cover up race condition
Okay so like, if we do this:

   room.assert('some fact')
   room.retract('some other fact')

We don't know what order the room object will send data out, and the
assert() or retract() could empty out this._messages before the second
call would complete sometimes.

Now we only remove confirmed messages that got sent, but we still are
doing a reassignment with this filter, so we might include already sent
messages. That is to say, we may sometimes double-send an assertion or
retraction, but at least we are not dropping any queued up messages.

I am not sure if this is fine, because I am not sure if
assertions are retractions are idempotent.

* A double assert should be idempotent
* A double retract should be idempotent
* Mixed asserts & retracts may not be idempotent

To be fair, that was always the case, to ensure a transaction use:

    room.assert().retract().assert()

There are probably other cases I am thinking of (specifically
subscriptions that have side effects) that might still fail.